% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_datatypes}
\alias{get_datatypes}
\title{returns the valid sample shape in order to choose a test (Internal Function)}
\usage{
get_datatypes(data = NULL, subsets = NULL)
}
\arguments{
\item{data}{The data to check (vector of integers).}
}
\value{
a list of the types of data
}
\description{
Checks the shape of the data and returns the global types
}
\keyword{internal}
