% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_test_from_string}
\alias{get_test_from_string}
\title{Internal: Returns the result of a statistical test based on a string identifier}
\usage{
get_test_from_string(test_object)
}
\arguments{
\item{test_object}{An object containing data, identifiers, datatypes, and test selection.}
}
\value{
The result of the selected statistical test. Typically, this is a test object with
        class `htest`, `aov`, `lm`, or similar.
}
\description{
This internal function selects and runs a statistical test using data from a test object,
based on the name of the test specified. It supports a wide variety of tests including
t-tests, chi-square tests, ANOVA, correlation tests, regression models, and more.
}
\keyword{internal}
