% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-picker.R
\name{pick_multiple_variable_test}
\alias{pick_multiple_variable_test}
\title{Pick the appropriate test for multiple variables (Internal Function)}
\usage{
pick_multiple_variable_test(test_object)
}
\arguments{
\item{test_object}{An object containing the data, types, and metadata needed for test selection.}
}
\value{
A character string with the name of the appropriate regression or classification model.
}
\description{
Pick the appropriate test for multiple variables (Internal Function)
}
\keyword{internal}
