% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-picker.R
\name{pick_one_variable_test}
\alias{pick_one_variable_test}
\title{Pick the appropriate test for one variable (Internal Function)}
\usage{
pick_one_variable_test(test_object)
}
\arguments{
\item{test_object}{An object containing the data, data types, and comparison value.}
}
\value{
A character string with the name of the appropriate one-sample statistical test.
}
\description{
Pick the appropriate test for one variable (Internal Function)
}
\keyword{internal}
