% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-picker.R
\name{pick_test}
\alias{pick_test}
\title{Check if a dataframe is parametric (Internal Function)
inst}
\usage{
pick_test(test_object)
}
\arguments{
\item{data}{The data to check (vector of integers).}
}
\value{
TRUE if data is normalized, FALSE otherwise.
}
\description{
Check if a dataframe is parametric (Internal Function)
inst
}
\keyword{internal}
