% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-picker.R
\name{pick_two_variable_test}
\alias{pick_two_variable_test}
\title{Pick the appropriate test for two variables (Internal Function)}
\usage{
pick_two_variable_test(test_object)
}
\arguments{
\item{test_object}{An object containing the data, types, and metadata needed for test selection.}
}
\value{
A character string with the name of the appropriate statistical test.
}
\description{
Pick the appropriate test for two variables (Internal Function)
}
\keyword{internal}
