% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-AutomatedTest.R
\name{AutomatedTest}
\alias{AutomatedTest}
\title{AutomatedTest class}
\description{
The AutomatedTest class is the class that represent a result of a statistical test. It contains all the attributes of the results such as the p-value, degrees of
freedom, and more.
}
\examples{
AutomatedTest$new(
  test = "Unpaired T-test",
)

}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AutomatedTest-new}{\code{AutomatedTest$new()}}
\item \href{#method-AutomatedTest-getDatatypes}{\code{AutomatedTest$getDatatypes()}}
\item \href{#method-AutomatedTest-setDatatypes}{\code{AutomatedTest$setDatatypes()}}
\item \href{#method-AutomatedTest-getParametric}{\code{AutomatedTest$getParametric()}}
\item \href{#method-AutomatedTest-setParametric}{\code{AutomatedTest$setParametric()}}
\item \href{#method-AutomatedTest-getTest}{\code{AutomatedTest$getTest()}}
\item \href{#method-AutomatedTest-setTest}{\code{AutomatedTest$setTest()}}
\item \href{#method-AutomatedTest-clone}{\code{AutomatedTest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutomatedTest-new"></a>}}
\if{latex}{\out{\hypertarget{method-AutomatedTest-new}{}}}
\subsection{Method \code{new()}}{
Initialize an instance of the AutomatedTest class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutomatedTest$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutomatedTest-getDatatypes"></a>}}
\if{latex}{\out{\hypertarget{method-AutomatedTest-getDatatypes}{}}}
\subsection{Method \code{getDatatypes()}}{
Get the datatypes of the features in the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutomatedTest$getDatatypes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of datatypes (e.g., Quantitative or Qualitative).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutomatedTest-setDatatypes"></a>}}
\if{latex}{\out{\hypertarget{method-AutomatedTest-setDatatypes}{}}}
\subsection{Method \code{setDatatypes()}}{
Set the datatypes for the features in the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutomatedTest$setDatatypes(dataTypes)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataTypes}}{A list of datatypes (e.g., Quantitative or Qualitative).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutomatedTest-getParametric"></a>}}
\if{latex}{\out{\hypertarget{method-AutomatedTest-getParametric}{}}}
\subsection{Method \code{getParametric()}}{
Get the parametric test results of the features and subsets
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutomatedTest$getParametric()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of parametric test results for all features and subsets.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutomatedTest-setParametric"></a>}}
\if{latex}{\out{\hypertarget{method-AutomatedTest-setParametric}{}}}
\subsection{Method \code{setParametric()}}{
Set the parametric test results
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutomatedTest$setParametric(parametric)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{parametric}}{A list that sets the parametric variable, including all parametric states of all features and subsets.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutomatedTest-getTest"></a>}}
\if{latex}{\out{\hypertarget{method-AutomatedTest-getTest}{}}}
\subsection{Method \code{getTest()}}{
Get the test results stored in the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutomatedTest$getTest()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of the test results.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutomatedTest-setTest"></a>}}
\if{latex}{\out{\hypertarget{method-AutomatedTest-setTest}{}}}
\subsection{Method \code{setTest()}}{
Set the test and its results
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutomatedTest$setTest(test)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{test}}{A list or object that sets the test and its results.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutomatedTest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AutomatedTest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutomatedTest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
