% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_parametric}
\alias{check_parametric}
\title{Check if a dataframe is parametric (Internal Function)
inst}
\usage{
check_parametric(data)
}
\arguments{
\item{data}{The data to check (vector of integers).}
}
\value{
TRUE if data is normalized, FALSE otherwise.
}
\description{
Check if a dataframe is parametric (Internal Function)
inst
}
\keyword{internal}
