% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_test_from_string}
\alias{get_test_from_string}
\title{Returns a statistical test function based on the name
of a test. (Internal Function)}
\usage{
get_test_from_string(test_object)
}
\arguments{
\item{name}{The name of a function}

\item{data}{The data to be used within the statistical test}
}
\value{
The corresponding function as a string, later to be ran.
}
\description{
Returns a statistical test function based on the name
of a test. (Internal Function)
}
\keyword{internal}
